Move view to views.py
