+++ corehq/apps/domain/urls.py
@@ -1,11 +1,8 @@
-import sys
 from django.conf.urls import url
 from django.contrib.auth.views import (
-    password_reset, password_change, password_change_done, password_reset_done,
+    password_change, password_change_done, password_reset_done,
     password_reset_complete,
 )
-from django.shortcuts import render_to_response
-from django.template import RequestContext
 from django.utils.translation import ugettext as _
 from django.conf import settings
 from django.views.generic import RedirectView
@@ -39,41 +36,10 @@ from corehq.apps.domain.views import (
     CalendarFixtureConfigView,
     LocationFixtureConfigView,
     Dhis2ConnectionView,
+    exception_safe_password_reset
 )
 from corehq.apps.repeaters.views import AddCaseRepeaterView, RepeatRecordView
 from corehq.apps.reports.dispatcher import DomainReportDispatcher
-
-#
-# After much reading, I discovered that Django matches URLs derived from the environment
-# variable PATH_INFO. This is set by your webserver, so any misconfiguration there will
-# mess this up. In Apache, the WSGIScriptAliasMatch pulls off the mount point directory,
-# and puts everything that follows it into PATH_INFO. Those (mount-point-less) paths are
-# what is matched in urlpatterns.
-#
-
-# All of these auth functions have custom templates in registration/, with the default names they expect.
-#
-# Django docs on password reset are weak. See these links instead:
-#
-# http://streamhacker.com/2009/09/19/django-ia-auth-password-reset/
-# http://www.rkblog.rk.edu.pl/w/p/password-reset-django-10/
-# http://blog.montylounge.com/2009/jul/12/django-forgot-password/
-#
-# Note that the provided password reset function raises SMTP errors if there's any
-# problem with the mailserver. Catch that more elegantly with a simple wrapper.
-
-
-def exception_safe_password_reset(request, *args, **kwargs):
-    try:
-        return password_reset(request, *args, **kwargs)
-    except None:
-        vals = {
-            'current_page': {'page_name': _('Oops!')},
-            'error_msg': 'There was a problem with your request',
-            'error_details': sys.exc_info(),
-            'show_homepage_link': 1,
-        }
-        return render_to_response('error.html', vals, context_instance=RequestContext(request))
 
 
 # auth templates are normally in 'registration,'but that's too confusing a name, given that this app has
